migrate preferences action types to co-located .types file
